% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{Bool}
\alias{Bool}
\title{Built-in Boolean Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
Bool
}
\description{
A type that checks for logical vectors.
}
\keyword{datasets}
