% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{DataFrame}
\alias{DataFrame}
\title{Built-in DataFrame Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
DataFrame
}
\description{
A type that checks for data.frame objects.
}
\keyword{datasets}
