% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{Double}
\alias{Double}
\title{Built-in Double Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
Double
}
\description{
A type that checks for double-precision numeric vectors.
}
\keyword{datasets}
