% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Enum}
\alias{Enum}
\title{Enum Type Factory}
\usage{
Enum(...)
}
\arguments{
\item{...}{Allowed scalar values or a single atomic vector of allowed values.}
}
\value{
A new sicher_type that checks all values belong to the enum.
}
\description{
Creates an enumeration type using regular function call syntax. The resulting type only
accepts atomic vectors whose elements all belong to the declared set of
allowed values.
}
\examples{
status \%:\% Enum(1, 2, 3) \%<-\% 2
colors \%:\% Enum("red", "green", "blue") \%<-\% c("red", "blue")

}
