% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{Function}
\alias{Function}
\title{Built-in Function Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
Function
}
\description{
A type that checks for function objects.
}
\keyword{datasets}
