% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Type-checked assignment operator}
\usage{
typed_annotation \%<-\% value
}
\arguments{
\item{typed_annotation}{Result of `\%:\%`.}

\item{value}{Value to assign.}
}
\value{
Invisibly returns the assigned value.
}
\description{
Completes the typed assignment started with `\%:\%`.
}
\examples{
\dontrun{
x \%:\% Integer \%<-\% 5L
y \%:\% Double \%<-\% 3.14
name \%:\% String \%<-\% "Bob"
flag \%:\% Bool \%<-\% TRUE
}

}
