% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{Integer}
\alias{Integer}
\title{Built-in Integer Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
Integer
}
\description{
A type that checks for integer vectors.
}
\keyword{datasets}
