% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{List}
\alias{List}
\title{Built-in List Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
List
}
\description{
A type that checks for list objects.
}
\keyword{datasets}
