% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{Null}
\alias{Null}
\title{Built-in Null Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
Null
}
\description{
A type that checks for NULL values.
}
\keyword{datasets}
