% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{Numeric}
\alias{Numeric}
\title{Built-in Numeric Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
Numeric
}
\description{
A type that checks for numeric vectors (integer or double).
}
\keyword{datasets}
