% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Optional}
\alias{Optional}
\title{Create an optional (nullable) type variant}
\usage{
Optional(type)
}
\arguments{
\item{type}{A sicher_type object}
}
\value{
A union type that includes Null
}
\description{
Creates a type that accepts NULL values in addition to the base type.
}
\examples{
middle_name \%:\% Optional(String) \%<-\% NULL
middle_name <- "Marie"  # Also OK

}
