% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{|.sicher_type}
\alias{|.sicher_type}
\title{Union Type Operator for sicher_type}
\usage{
\method{|}{sicher_type}(type1, type2)
}
\arguments{
\item{type1}{First type (sicher_type object)}

\item{type2}{Second type (sicher_type object)}
}
\value{
A union type (sicher_union object)
}
\description{
S3 method for the `|` operator to create union types.
}
