% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{|.sicher_union}
\alias{|.sicher_union}
\title{Union Type Operator for sicher_union}
\usage{
\method{|}{sicher_union}(type1, type2)
}
\arguments{
\item{type1}{First type (sicher_union object)}

\item{type2}{Second type}
}
\value{
A union type (sicher_union object)
}
\description{
S3 method for the `|` operator to extend union types.
}
