% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.sicher_type}
\alias{print.sicher_type}
\title{Print method for sicher_type}
\usage{
\method{print}{sicher_type}(x, ...)
}
\arguments{
\item{x}{A sicher_type object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for sicher_type
}
