% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.sicher_typed_annotation}
\alias{print.sicher_typed_annotation}
\title{Print method for sicher_typed_annotation}
\usage{
\method{print}{sicher_typed_annotation}(x, ...)
}
\arguments{
\item{x}{A sicher_typed_annotation object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for sicher_typed_annotation
}
