% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.sicher_union}
\alias{print.sicher_union}
\title{Print method for sicher_union}
\usage{
\method{print}{sicher_union}(x, ...)
}
\arguments{
\item{x}{A sicher_union object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for sicher_union
}
