% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Readonly}
\alias{Readonly}
\title{Create a readonly type variant}
\usage{
Readonly(type)
}
\arguments{
\item{type}{A sicher_type object}
}
\value{
A readonly type modifier
}
\description{
Creates a type that prevents reassignment after initial value is set.
}
\examples{
\dontrun{
PI \%:\% Readonly(Double) \%<-\% 3.14159
# PI <- 3.0  # Error: cannot reassign readonly
}

}
