% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Scalar}
\alias{Scalar}
\title{Create a scalar (length-1) type variant}
\usage{
Scalar(type)
}
\arguments{
\item{type}{A sicher_type object}
}
\value{
A new sicher_type that checks for length 1
}
\description{
Creates a type that only accepts single values (vectors of length 1).
}
\examples{
\dontrun{
age \%:\% Scalar(Integer) \%<-\% 30L
# age <- c(30L, 40L)  # Error: not scalar
}

}
