% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{String}
\alias{String}
\title{Built-in String Type}
\format{
An object of class \code{sicher_type} of length 2.
}
\usage{
String
}
\description{
A type that checks for character vectors.
}
\keyword{datasets}
