% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{[.sicher_type}
\alias{[.sicher_type}
\title{Vector Size Operator for sicher_type}
\usage{
\method{[}{sicher_type}(type, size)
}
\arguments{
\item{type}{A sicher_type object}

\item{size}{The required length (non-negative integer)}
}
\value{
A new sicher_type that checks for the specified length
}
\description{
Creates a type that checks for a specific vector length.
}
\examples{
\dontrun{
vec \%:\% Numeric[3] \%<-\% c(1, 2, 3)
# vec <- c(1, 2)  # Error: wrong length
# vec <- c("a", "b", "c")  # Error: wrong type
}

}
