% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{union-operator}
\alias{union-operator}
\alias{|.sicher_type}
\alias{|.sicher_union}
\title{Union Type Operator}
\usage{
\method{|}{sicher_type}(type1, type2)

\method{|}{sicher_union}(type1, type2)
}
\arguments{
\item{type1}{First type (sicher_type or sicher_union object)}

\item{type2}{Second type (sicher_type or sicher_union object)}
}
\value{
A union type (sicher_union object)
}
\description{
S3 methods for the `|` operator to create union types.
}
